refactor: 大规模代码重构和构建系统优化
- 移除旧的 CMake 脚本 (EnsureNlohmannJson.cmake, EnsureSpdlog.cmake, FindScoopLua.cmake)
- 新增 CLI 模块 (cli.hpp, cli.cpp) 及相关测试
- 优化 runtime 命令分发逻辑，提升代码可读性
- 改进 macOS 平台代码 (cocoa_screen, mac_automation 等)
- 增强 security 和 screen 模块功能
- 统一各模块 CMakeLists.txt 构建配置
- 更新 vcpkg 依赖配置

+++ lib/wingman/src/performance.cpp
@@ -300,7 +300,7 @@ std::vector<Point> PerformanceManager::parallelFindColors(
         for (int x = 0; x < width; ++x) {
             if (maskRow[x]) {
                 results.emplace_back(region.x + x, region.y + y);
-                if (maxCount > 0 && results.size() >= maxCount) {
+                if (maxCount > 0 && results.size() >= static_cast<size_t>(maxCount)) {
                     return results;
                 }
             }
